Linked to homescreen after logging in. Moved homescreen to screens folder.
